Add files via upload
The Amazon rainforest, considered the "green lung of the planet," hosts unparalleled biodiversity and plays a crucial role in global climate regulation. However, in recent years, it has faced a worrying deforestation process that threatens its ecological balance. The loss of biodiversity, increased CO2 emissions, and disruption of the water cycle are just some of the negative effects of deforestation that harm all living beings. This project aims to contribute to the automatic analysis of deforestation in the Amazon rainforest by using image segmentation techniques. The main goal is to develop an automated method to calculate the deforested area in a specific region of the Jamanxim National Forest in Brazil, using various satellite images.

In this project, different image processing and segmentation techniques will be applied to analyze the forest's deforestation. The selected images for this study cover a period of 16 years, from 2000 to 2016. The image from the year 2000 serves as the initial reference point, showing a landscape mostly covered by lush vegetation with early signs of deforestation. In contrast, the image captured 16 years later reveals substantial changes in the landscape with clear signs of deforestation.